move layout fixtures into conftest.py
